py/requirements.txt
Flask==2.1.0
openai
python-dotenv